requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
jinja2
pydantic
pydantic-settings
